Rename the index.js file
