+++ src/redux/store.js
@@ -1,6 +1,6 @@
 import { applyMiddleware, createStore } from 'redux';
 import thunk from 'redux-thunk';
-import rootReducer from './reducers/rootReducer';
+import rootReducer from './reducers/index';
 
 const store = createStore(rootReducer, applyMiddleware(thunk));
